level-3 rename src/stylesheets/*.js to src/javascripts/*_style.js
